Helper function cleanup
Inner functions that were in `find_plethysms` and `young_symmetrizer`
were turned into private helper functions. There is now a list of these
helper functions in the beginning of the script. They all come with `_`
to show they are supposed to be private. The idea is that the user does
not really need access to these functions; or at least, the authors
believe these functions fall outside the scope of what a user should get
out of the code. This is a first step to splitting the current script
into modules.