Symmetric group overhaul. (#2)
* Symmetric group overhaul.

* Helper function cleanup

Inner functions that were in `find_plethysms` and `young_symmetrizer`
were turned into private helper functions. There is now a list of these
helper functions in the beginning of the script. They all come with `_`
to show they are supposed to be private. The idea is that the user does
not really need access to these functions; or at least, the authors
believe these functions fall outside the scope of what a user should get
out of the code. This is a first step to splitting the current script
into modules.

* fixed reference

Unfortunately, I had linked an older reference for the algorithm we use
to build Young symmetrizers. I have fixed this to reference the actual
algorithm we use.

* Split up script into package, imports still broken.

* Fixed imports.

* Wrapped CGC_Data creation inside a function.

* Fixed CGC directory location

* Fixed missing import in plethysm_utils

* Added demo script

---------

Co-authored-by: Jason K. Elhaderi <jasonke2@illinois.edu>